float 16 to float 8

+++ mask_detector/train_mask_detector.py
@@ -52,7 +52,7 @@ lb = LabelBinarizer()
 labels = lb.fit_transform(labels)
 labels = to_categorical(labels)
 
-data = np.array(data, dtype="float16")#"float32"
+data = np.array(data, dtype="float8")#"float32"
 labels = np.array(labels)
 
 (trainX, testX, trainY, testY) = train_test_split(data, labels,
